added lives instead of news
