+++ app/assets/stylesheets/lives.scss
@@ -1,44 +1,187 @@
+@import 'reset';
+@import 'mixin';
+@import 'fonts';
+
 .lives{
-  margin: 15px auto;
   width: 100%;
-  color: #ffffff;
-  h5{
-    font-size: 20px;
-    margin: 5px;
-    justify-content: start;
+  padding: 20px 10px;
+  margin: 0 auto;
+  @include pc {
+    max-width: 1000px;
   }
-  &_info{
-    margin: 10px 25%;
-    &_title{
-      display: flex;
-      .date{
-        font-size: 15px;
-        width: 150px;
+  &-tabs{
+    display: flex;
+    margin-top: 10px;
+    z-index: 20;
+    div{
+      width: 120px;
+      transition: transform 0.2s;
+      background-color: #111111;
+      border: 1px solid #ffffff;
+      padding: 0 10px;
+      margin-right: 5px;
+      &:hover{
+        transform: scale(1.1);
+        cursor: pointer;
       }
-      h3, a{
-        color: #ffffff;
-        font-size: 20px;
-        margin-left: 25px;
-      }
-    }
-    &_sub{
-      display: flex;
-      &_img{
-        height: 100px;
-        width: 150px;
-        font-size: 75px;
-        font-family: 'Impact';
-        text-align: center;
-      }
-      &_text{
-        margin: auto;
-        // justify-content: center;
-        // align-items: center;
+      &.selected{
+        background-color: #ffffff;
+        color: #111111;
       }
     }
   }
-  .time{
-    font-size: 15px;
-    margin-left: auto;
+  &-future, &-past{
+    height: 100vh;
+    overflow: scroll;
+  }
+  .list{
+    &-schedules{
+      @include pc {
+        display: flex;
+        flex-wrap: wrap;
+        visibility: hidden;
+        .list-schedules-item{
+          visibility: visible;
+          position: relative;
+          &:before{
+            content: '';
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-color: transparent;
+            pointer-events: none;
+            z-index: 999;
+            transition: background-color .5s;
+          }
+        }
+        &:hover{
+          .list-schedules-item{
+            &:not(:hover){
+              &:before{
+                pointer-events: none;
+                background-color: rgba(0,0,0,0.5);
+              }
+            }
+          }
+        }
+      }
+      &-item{
+        border: 1px solid #fff;
+        @include pc {
+          padding: 20px;
+          width: calc((100% - 20px)/2);
+          &:nth-child(even){
+            margin-left: 20px;
+          }
+          &:nth-child(n+3){
+            margin-top: 20px;
+          }
+        }
+        @include sp {
+          padding: 10px;
+          &:not(:first-child){
+            margin-top: 10px;
+          }
+        }
+        .item{
+          &-ttl{
+            font-weight: bold;
+            text-decoration: underline;
+            margin-bottom: 2rem;
+            @include pc {
+              font-size: 1.8rem;
+            }
+            @include sp {
+              font-size: calcvw(32);
+            }
+            white-space: nowrap;
+            text-overflow: ellipsis;
+            overflow: hidden;
+            a{
+              color: #fff;
+            }
+          }
+          &-date{
+            margin-bottom: 2rem;
+            @include pc {
+              font-size: 1.4rem;
+            }
+            @include sp {
+              font-size: calcvw(22);
+            }
+            white-space: nowrap;
+            text-overflow: ellipsis;
+            overflow: hidden;
+          }
+          &-description{
+            white-space: nowrap;
+            text-overflow: ellipsis;
+            overflow: hidden;
+            margin-bottom: 2rem;
+            @include pc {
+              font-size: 1.6rem;
+            }
+            @include sp {
+              font-size: calcvw(24);
+            }
+          }
+          &-thumbnails{
+            display: flex;
+            justify-content: center;
+            .thumbnail{
+              @include pc {
+                width: calc((100% - 20px)/2);
+                &:nth-child(2){
+                  margin-left: 20px;
+                }
+              }
+              @include sp {
+                width: calc((100% - 10px)/2);
+              &:nth-child(2){
+                margin-left: 10px;
+              }
+              }
+              overflow: hidden;
+              position: relative;
+              margin-bottom: 2rem;
+              &:before{
+                content: '';
+                display: block;
+                padding-top: 100%;
+              }
+              img{
+                display: block;
+                width: 100%;
+                position: absolute;
+                top: 0;
+                left: 0;
+              }
+            }
+            .no_image{
+              height: 300px;
+              width: 300px;
+              font: 50px impact;
+            }
+          }
+          &-details{
+            &-time{
+              display: flex;
+              p{
+                margin-right: 10px;
+              }
+            }
+          }
+        }
+        &:hover{
+          .item-btns{
+            @include pc {
+              opacity: 1;
+            }
+          }
+        }
+      }
+    }
   }
 }
