+++ app/assets/stylesheets/main.scss
