+++ app/views/home/main.html.haml
@@ -42,15 +42,15 @@
         %p テキストテキストテキスト
       .news
         %h2 News
-        %p.paratext
-          Lorem ipsum, dolor sit amet consectetur adipisicing elit. Exercitationem hic eligendi nesciunt animi omnis! Officiis, est accusamus? Et explicabo repellendus impedit optio eos vero tempore earum. Voluptates consequatur voluptatem repellendus.
-        %p.paratext
-          Lorem ipsum, dolor sit amet consectetur adipisicing elit. Exercitationem hic eligendi nesciunt animi omnis! Officiis, est accusamus? Et explicabo repellendus impedit optio eos vero tempore earum. Voluptates consequatur voluptatem repellendus.
-        %p.paratext
-          Lorem ipsum, dolor sit amet consectetur adipisicing elit. Exercitationem hic eligendi nesciunt animi omnis! Officiis, est accusamus? Et explicabo repellendus impedit optio eos vero tempore earum. Voluptates consequatur voluptatem repellendus.
-        %p.paratext
-          Lorem ipsum, dolor sit amet consectetur adipisicing elit. Exercitationem hic eligendi nesciunt animi omnis! Officiis, est accusamus? Et explicabo repellendus impedit optio eos vero tempore earum. Voluptates consequatur voluptatem repellendus.
-      = render partial: 'home/lives'
+        .lives
+          %h2 Live Information
+          .lives-tabs
+            .future.selected UPCOMING
+            .past PAST
+          .lives-future
+            = render partial: 'home/lives', locals:{lives: @future_lives}
+          .lives-past.hidden
+            = render partial: 'home/lives', locals:{lives: @past_lives}
       .about
         %h2 About
         %p テキストテキストテキスト
